// src/api/models/api_PatchUserVerseRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type api_PatchUserVerseRequest = {
    lastReviewedAt?: string;
    lastTrainingModeId?: number;
    masteryLevel?: number;
    nextReviewAt?: string;
    repetitions?: number;
    reviewLapseStreak?: number;
    status?: string;
};